requirements.txt
guessit
requests
pyperclip
Pillow
bencode.py